djangoooo/ourWebsite/requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
Automat==0.8.0
automium==0.2.6
automium-web==0.1.1
beautifulsoup4==4.9.0
bson==0.5.8
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
dataclasses==0.6
defusedxml==0.6.0
distlib==0.3.0
Django==2.2.12
django-allauth==0.41.0
django-bootstrap4==1.1.1
django-cors-headers==3.2.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
djongo==1.3.2
filelock==3.0.12
Flask==1.1.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.6.0
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
legacy==0.1.6
lxml==4.5.0
MarkupSafe==1.1.1
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.2.0.34
parsel==1.5.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==2.0.0
pymongo==3.10.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
queuelib==1.5.0
requests==2.23.0
requests-oauthlib==1.3.0
Scrapy==1.8.0
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.2.4
Twisted==19.10.0
urllib3==1.25.9
views==0.3
virtualenv==20.0.21
w3lib==1.21.0
Werkzeug==1.0.1
zipp==3.1.0
zope.interface==4.7.1
